requirements.txt
black==21.9b0
Flask==1.1.2
Flask-Session==0.4.0
flake8==3.9.2
requests==2.26.0
spotipy==2.19.0
python-dotenv==0.19.0
firebase-admin==5.0.3
frozendict==2.0.6
libsass==0.21.0